Use srand to seed the MCMC instead of step # + replicate #.
Fixes issue where the same seed gets set multiple times. For example,
seed=10 occurs at (step 1, replicate 9), (step 2, replicate 8), (step 3,
replicate 7), etc.

Now, an initial seed for the PGS rng is set using rand in Sim::run(),
and the seed for that one is set as a hyperparameter in the config file.
(Defaults to 1.)