Implement a locally-balanced MC sampler.
See https://arxiv.org/pdf/1711.07424 for details.

Rough first pass at it. Not efficient and very slow...

The issue that computing the local environment before sampling a
proposal is very expensive, and it only beats out the standard random
uniform proposal in instances where the underlying distribution is very
'rough.'

The issue is that the initial parameters (h and J) are pretty flat at
the very start and 'roughen up' as the inference loop proceeds. Using
this sampler is a very bad idea at the start of the loop. When (or if)
it becomes better at some point remains to be determined...