Enable bmDCA to restart stopped runs.
When starting bmDCA, the program will look inside the destination
directory (-d flag) and search for the most recently generated
parameters files. These will then be loaded into the inference model.
Additionally, the bmdca_run.log file will be parsed to reset the burn-in
times and 'spend' random numbers in the RNG in order for runs to be
reproducible. By that, I meant that runs completed without stopping will
produce _exactly_ the same output as those that stopped and started
again.

Additional behaviors added as a side effeect:
1. The parameters for the current _and_ previous step get saved every
   save_parameters number of steps. This is because the updates to the
   parmaeters, gradients, and learning rates depend on both the values
   from the current and previous step.
2. Files in the destination directory will be deleted if the values in
   the bmdca.conf file are changed between stopping and starting.
3. The bmdca_run.log will get duplicates of steps if bmDCA is stopped
   while writing files. This is because the run log gets written first.
4. Idential output for stopped and started runs vs continuous ones
   requires output files be stored in binary format. There isn't really
   a one-to-one mapping between ASCII and doubles, so one cannot truly
   be guaranteed identical output if parameters are saved in text format
   and then reloaded when restarting bmDCA.